requirements.txt
amqp==2.1.1
anyjson==0.3.3
awesome-slugify==1.6.5
Babel==2.3.4
beautifulsoup4==4.5.1
billiard==3.5.0.2
celery==4.0.0
decorator==4.0.10
defusedxml==0.4.1
Django==1.9.11
django-appconf==1.0.2
django-async-messages==0.3.1
django-autoslug==1.9.3
django-bootstrap3==7.0.1
django-braces==1.9.0
django-celery-email==1.1.5
django-ckeditor==5.1.1
django-countries==4.0
django-crispy-forms==1.6.0
django-debug-panel==0.8.3
django-debug-toolbar==1.6
django-grappelli==2.8.3
django-imagekit==3.3
django-memoize==2.0.0
django-model-utils==2.5.2
django-money==0.9.1
django-recaptcha==1.0.5
django-redis==4.4.4
Django-Select2==5.8.9
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
Faker==0.7.3
flower==0.9.1
gunicorn==19.6.0
html5lib==0.999999999
ipython==5.1.0
ipython-genutils==0.1.0
kombu==4.0.0
lxml==3.6.4
oauthlib==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==1.1.13
Pillow==3.3.1
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
py-moneyed==0.6.0
pycountry==1.20
Pygments==2.1.3
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2016.6.1
redis==2.10.5
regex==2016.8.27
requests==2.11.1
requests-oauthlib==0.6.2
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
tornado==4.2
traitlets==4.3.0
Unidecode==0.4.19
URLObject==2.4.2
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5
